rashad - AdminActivity works, NavigationBar need to be activated

+++ app/src/main/java/com/example/databases/AdminFacade.java
@@ -2,10 +2,11 @@ package com.example.databases;
 
 import android.content.Context;
 
+import java.io.Serializable;
 import java.text.ParseException;
 import java.util.ArrayList;
 
-public class AdminFacade extends  ClientFacade {
+public class AdminFacade extends  ClientFacade implements Serializable {
 
     public AdminFacade(Context context) {
         super(context);
@@ -27,8 +28,10 @@ public class AdminFacade extends  ClientFacade {
         Company existingCompany = companiesDAO.getOneCompany(company.getId());
 
         if (existingCompany != null) {
-            existingCompany.setId(company.getId());
+            //existingCompany.setId(company.getId());
             existingCompany.setName(company.getName());
+            existingCompany.setEmail(company.getEmail());
+            existingCompany.setPassword(company.getPassword());
             companiesDAO.updateCompany(existingCompany);
         } else {
             throw new myException("Company not found for the given ID.");
@@ -42,8 +45,10 @@ public class AdminFacade extends  ClientFacade {
         if (companyExists != null) {
             try {
                 ArrayList<Coupon> companyCoupons = company.getCoupons();
-                for (Coupon coupon : companyCoupons)
-                    couponsDAO.deleteCoupon(coupon);
+                if(companyCoupons!=null) { /// without this app will crash!
+                    for (Coupon coupon : companyCoupons)
+                        couponsDAO.deleteCoupon(coupon);
+                }
                 companiesDAO.deleteCompany(companyIdToDelete);
             } catch (myException e) {
                 throw new myException("Error deleting the company.");
@@ -89,7 +94,9 @@ public class AdminFacade extends  ClientFacade {
         if (customer != null) {
             try {
                 ArrayList<Coupon> customerCoupons = customer.getCoupons();
-                customerCoupons.clear();
+                if(customerCoupons!=null){
+                    customerCoupons.clear();
+                }
                 customersDAO.deleteCustomer(customerID);
             } catch (myException e) {
                 throw new myException("Error deleting the customer.");
@@ -99,8 +106,11 @@ public class AdminFacade extends  ClientFacade {
         }
     }
 
-    public ArrayList<Customer> gelAllCustomers() {
+    public ArrayList<Customer> getAllCustomers() {
         return customersDAO.getAllCustomers();
+    }
+    public Customer getOneCustomer(int customerID) {
+        return customersDAO.getOneCustomer(customerID);
     }
     @Override
     public boolean login(String email, String password) {
